requirements.txt
gluoncv
keras
numpy
mxnet
pillow